refactor : 세팅 페이지 변경
