+++ Android/app/src/main/java/com/booktree/ui/setting/SettingFragment.java
@@ -1,4 +1,4 @@
-package com.booktree.ui.notifications;
+package com.booktree.ui.setting;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,20 +10,19 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
-import com.booktree.R;
-import com.booktree.databinding.FragmentNotificationsBinding;
+import com.booktree.databinding.FragmentSettingBinding;
 
-public class NotificationsFragment extends Fragment {
+public class SettingFragment extends Fragment {
 
-  private NotificationsViewModel notificationsViewModel;
-  private FragmentNotificationsBinding binding;
+  private SettingViewModel notificationsViewModel;
+  private FragmentSettingBinding binding;
 
   public View onCreateView(@NonNull LayoutInflater inflater,
       ViewGroup container, Bundle savedInstanceState) {
     notificationsViewModel =
-        new ViewModelProvider(this).get(NotificationsViewModel.class);
+        new ViewModelProvider(this).get(SettingViewModel.class);
 
-    binding = FragmentNotificationsBinding.inflate(inflater, container, false);
+    binding = FragmentSettingBinding.inflate(inflater, container, false);
     View root = binding.getRoot();
 
     final TextView textView = binding.textNotifications;
