+++ Android/app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.notifications.NotificationsFragment">
+  tools:context=".ui.setting.SettingFragment">
 
   <TextView
     android:id="@+id/text_notifications"
